% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendedCat.R
\docType{methods}
\name{appendedCat}
\alias{appendedCat}
\alias{appendedCat,character,character-method}
\alias{appendedCat,character,missing-method}
\title{Prints first text file and appends into second file}
\usage{
appendedCat(firstText, secondText)

\S4method{appendedCat}{character,character}(firstText, secondText)

\S4method{appendedCat}{character,missing}(firstText)
}
\arguments{
\item{firstText}{character, First text.}

\item{secondText}{character, Second text.}
}
\value{
Appended text file.
}
\description{
This function is a generic function to print the first text and appends into second
             file if it exists.
}
\author{
Yong Luo
}
