% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiTreeCompiler.R
\docType{methods}
\name{auxiTreeCompiler}
\alias{auxiTreeCompiler}
\alias{auxiTreeCompiler,data.table,data.table,data.table-method}
\title{Derive volume components for H-enhanced and non-enhanced trees-VRI specific}
\usage{
auxiTreeCompiler(fullMeasuredTrees, auxiTrees, clusterPlotHeader)


  \S4method{auxiTreeCompiler}{data.table,data.table,data.table}(fullMeasuredTrees,
  auxiTrees, clusterPlotHeader)
}
\arguments{
\item{fullMeasuredTrees}{Compiled tree-level data in vi_c, which contains full measured trees, enhanced trees
and H-enhanced trees. This data is output of \code{\link{DWBCompiler}}}

\item{auxiTrees}{data.table, Non-enhanced trees in anxilirary plots, however, it may have enhanced trees and H-enhanced trees.
An output from \code{\link{VRIInit_auxTree}}.}

\item{clusterPlotHeader}{data.table, Cluster and plot-level information. An output of \code{\link{VRIInit_clusterplot}}.}
}
\value{
A list of four tables: 1. fullenhancedtrees: full and enhanced trees; 
        2. HnonenhancedTrees: Height enhanced and non-enhanced trees;
        3. regression table; 4. ratio table.
}
\description{
Estimates volume components for H-enhanced and non-enhanced trees using regression and ratio methods.
             For H-enhanced trees, the whole stem volume and gross merchantable volume are already calculated directly using
             taper equations; and rest of volume components will be calculated using ratio method in this function.
             For non-enhanced trees, the whole stem volume is derived using regression equation between basal area
             and whole stem volume and the rest of volume components will be computed using ratio method in this function.
             The function is part of \code{vol_ha_2017.sas}.
}
\author{
Yong Luo
}
