% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boredAgeCalculator_Crted.R
\docType{methods}
\name{boredAgeCalculator_Crted}
\alias{boredAgeCalculator_Crted}
\alias{boredAgeCalculator_Crted,numeric,numeric,numeric,character,character,character,character-method}
\alias{boredAgeCalculator_Crted,numeric,numeric,numeric,character,character,character,missing-method}
\title{Derive age at breast height}
\usage{
boredAgeCalculator_Crted(boredAge, boredHeight, treeHeight, species, ICRegion,
  siteToolsDLLPath, sasExePath)


  \S4method{boredAgeCalculator_Crted}{numeric,numeric,numeric,character,character,character,character}(boredAge,
  boredHeight, treeHeight, species, ICRegion, siteToolsDLLPath, sasExePath)


  \S4method{boredAgeCalculator_Crted}{numeric,numeric,numeric,character,character,character,missing}(boredAge,
  boredHeight, treeHeight, species, ICRegion, siteToolsDLLPath)
}
\arguments{
\item{boredAge}{numeric, Age at bored height}

\item{boredHeight}{numeric, Height at the bore core.}

\item{treeHeight}{numeric, Total tree height}

\item{species}{character, Species code, must be consistent with the species code in site tools.
Can be derived using \code{\link{siteToolsSpeciesConvertor}}.}

\item{ICRegion}{character, Must be either \code{I} (interior) and \code{C} (coastal). 
IC regions can be derived using \code{\link{BEC2IC}} function.}

\item{siteToolsDLLPath}{character, Path to \code{SINDEX33.DLL}}

\item{sasExePath}{character, Path to sas executable, i.e., \code{sas.exe}. If missing, the function takes
\code{C:/Program Files/SASHome/x86/SASFoundation/9.3} as default. However, it will cause crush
if sas executable is not located in default path.}
}
\value{
Tree age at breast height (1.3 m). For the estimated age <= 0.1, the bored age will be used as breast age and 
             a warning message will be given.
}
\description{
This function uses site tools to derive age at breast height for the bored age that is not taken at 
             breast height, i.e., \code{1.3 m}. The process was documented in BC VRI Sample Data Compilation Process.
}
\author{
Yong Luo
}
