% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilerOutputSetup.R
\docType{methods}
\name{compilerOutputSetup}
\alias{compilerOutputSetup}
\alias{compilerOutputSetup,character-method}
\alias{compilerOutputSetup,missing-method}
\title{Setup an output path of the compiler}
\usage{
compilerOutputSetup(outputPath)

\S4method{compilerOutputSetup}{character}(outputPath)

\S4method{compilerOutputSetup}{missing}()
}
\arguments{
\item{outputPath}{character, Specifies an output path.
If missing, the current work directory will be used.}
}
\value{
An output path that will be used to store your compiled data
}
\description{
This function does two things: 1. create a folder that will store compiled data;
             2. return a path that directs the compiled folder.
}
\note{
Could overwrite the existing output folder, depending on user's choise, i.e., yes or no.
}
\author{
Yong Luo
}
