% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteAgeCompiler.R
\docType{methods}
\name{siteAgeCompiler}
\alias{siteAgeCompiler}
\alias{siteAgeCompiler,data.table,character,character-method}
\alias{siteAgeCompiler,data.table,character,missing-method}
\title{Compile breast age, total age, and site index where possible-VRI specific}
\usage{
siteAgeCompiler(siteAgeData, siteToolsDLLPath, sasExePath)

\S4method{siteAgeCompiler}{data.table,character,character}(siteAgeData,
  siteToolsDLLPath, sasExePath)

\S4method{siteAgeCompiler}{data.table,character,missing}(siteAgeData,
  siteToolsDLLPath)
}
\arguments{
\item{siteAgeData}{data.table, Site age data with plot header information.
An output from \code{\link{VRIInit_siteTree}} function.}

\item{siteToolsDLLPath}{character, Path to \code{SINDEX33.DLL}.}

\item{sasExePath}{character, Path to sas executable, i.e., \code{sas.exe}. If missing, the function takes
\code{C:/Program Files/SASHome/x86/SASFoundation/9.3} as default. However, it will cause crush
if sas executable is not located in default path.}
}
\value{
A data table and a log file.
}
\description{
This function takes site age tree data ie., \code{vi_h}, an output of \code{\link{VRIInit_siteTree}} to
             compute the breast height age, total age, and site index where possible. This function is
             equivalent to site_age.sas. The function heavily depends on site tools program.
}
\author{
Yong Luo
}
