% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiteTools_HTBoredAge2SI.R
\docType{methods}
\name{SiteTools_HTBoredAge2SI}
\alias{SiteTools_HTBoredAge2SI}
\alias{SiteTools_HTBoredAge2SI,numeric,numeric,character,character,numeric,numeric,character,character-method}
\title{Calculate site index using site tools}
\usage{
SiteTools_HTBoredAge2SI(boredAge, height, species, ICRegion, ageType,
  estimateMethod, siteToolsDLLPath, sasExePath)


  \S4method{SiteTools_HTBoredAge2SI}{numeric,numeric,character,character,numeric,numeric,character,character}(boredAge,
  height, species, ICRegion, ageType, estimateMethod, siteToolsDLLPath,
  sasExePath)
}
\arguments{
\item{boredAge}{numeric, Age at bored height.}

\item{height}{numeric, Total tree height.}

\item{species}{character, Species code, must be consistent with the species code in site tools, which can be converted 
from the original species code by using \code{\link{siteToolsSpeciesConvertor}}.}

\item{ICRegion}{character, Must be either \code{I} (interior) and \code{C} (coastal). 
IC regions can be derived using \code{\link{BEC2IC}}.}

\item{ageType}{numeric, Must be either \code{0} or \code{1}. \code{0} stands for total age, for which site index is 
calculated for 50 years of total tree age. While \code{1} stands for breast height age, for which 
site index is calculated for 50 year old at breast height.}

\item{estimateMethod}{numeric, Defines how the site tools estimate site index. Valued as \code{0} and \code{1}, 
\code{0} is interative and while \code{1} is directive. Default is \code{1}, which is directive.}

\item{siteToolsDLLPath}{character, Path to \code{SINDEX33.DLL}}

\item{sasExePath}{character, Path to sas executable, i.e., \code{sas.exe}.}
}
\value{
Site index
}
\description{
This function calculates site index based on bored age (\code{boredAge}), tree height (\code{height}),
             species (\code{species}) and region (\code{ICRegion}) using site tools program. This function
             is equivalent to sindex_httoage.sas.
}
\author{
Yong Luo
}
