% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiteTools_Y2BH.R
\docType{methods}
\name{SiteTools_Y2BH}
\alias{SiteTools_Y2BH}
\alias{SiteTools_Y2BH,character,character,numeric,character,character-method}
\alias{SiteTools_Y2BH,character,character,numeric,character,missing-method}
\title{Derive years to breast height using site tools}
\usage{
SiteTools_Y2BH(species, ICRegion, siteIndex, siteToolsDLLPath, sasExePath)


  \S4method{SiteTools_Y2BH}{character,character,numeric,character,character}(species,
  ICRegion, siteIndex, siteToolsDLLPath, sasExePath)


  \S4method{SiteTools_Y2BH}{character,character,numeric,character,missing}(species,
  ICRegion, siteIndex, siteToolsDLLPath)
}
\arguments{
\item{species}{character, Species code, must be consistent with the species code in site tools, which can be converted 
from the original species code by using \code{\link{siteToolsSpeciesConvertor}}.}

\item{ICRegion}{character, Must be either \code{I} (interior) and \code{C} (coastal). 
IC regions can be derived using \code{\link{BEC2IC}} function.}

\item{siteIndex}{numeric, Site index. Defined as tree height at 50 years old.}

\item{siteToolsDLLPath}{character, Path to \code{SINDEX33.DLL}}

\item{sasExePath}{character, Path to sas executable, i.e., \code{sas.exe}. If missing, the function takes
\code{C:/Program Files/SASHome/x86/SASFoundation/9.3} as default. However, it will cause crush
if sas executable is not located in default path.}
}
\value{
Years to breast height
}
\description{
Derive years to breast height based on species (\code{species}), region (\code{ICRegion}) and site index (\code{siteIndex})
             using site tools. This function is equivalent to \code{sindex_httoage.sas}.
}
\author{
Yong Luo
}
