% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callSiteTools.R
\docType{methods}
\name{ST_CurveName}
\alias{ST_CurveName}
\alias{ST_CurveName,numeric,character,character-method}
\alias{ST_CurveName,numeric,character,missing-method}
\title{Derive site index curve name based on site index curve reference}
\usage{
ST_CurveName(SICurveRef, siteToolsDLLPath, sasExePath)

\S4method{ST_CurveName}{numeric,character,character}(SICurveRef,
  siteToolsDLLPath, sasExePath)

\S4method{ST_CurveName}{numeric,character,missing}(SICurveRef, siteToolsDLLPath)
}
\arguments{
\item{SICurveRef}{numeric, Site index curve reference in site tools. It can be derived using \code{ST_DefCurve} function.}

\item{siteToolsDLLPath}{character, Path to \code{SINDEX33.DLL}}

\item{sasExePath}{character, Path to sas executable, i.e., \code{sas.exe}. If missing, the function takes
\code{C:/Program Files/SASHome/x86/SASFoundation/9.3} as default. However, it will cause crush
if sas executable does not located in default folder.}
}
\value{
site index curve name
}
\description{
Derive site index curve name based on site index curve reference (\code{SICurveRef}) in site tools.
}
\author{
Yong Luo
}
