% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callSiteTools.R
\docType{methods}
\name{ST_DefCurve}
\alias{ST_DefCurve}
\alias{ST_DefCurve,numeric,character,character-method}
\alias{ST_DefCurve,numeric,character,missing-method}
\title{Assign site index curve reference in site tools}
\usage{
ST_DefCurve(siteIndexRef, siteToolsDLLPath, sasExePath)

\S4method{ST_DefCurve}{numeric,character,character}(siteIndexRef,
  siteToolsDLLPath, sasExePath)

\S4method{ST_DefCurve}{numeric,character,missing}(siteIndexRef,
  siteToolsDLLPath)
}
\arguments{
\item{siteIndexRef}{numeric, Site index reference. It can be derived using \code{\link{ST_SpecRemap}} function.}

\item{siteToolsDLLPath}{character, Path to \code{SINDEX33.DLL}}

\item{sasExePath}{character, Path to sas executable, i.e., \code{sas.exe}. If missing, the function takes
\code{C:/Program Files/SASHome/x86/SASFoundation/9.3} as default. However, it will cause crush
if sas executable does not located in default folder.}
}
\value{
Site index curve reference that can be recognized by site tools
}
\description{
Assign site index curve reference that can be recognized by site tools for a given site index reference.
             The site index can be generated using \code{\link{ST_SpecRemap}} function.
}
\seealso{
\code{\link{ST_SpecRemap}}, \code{\link{ST_DefGICurve}}, \code{\link{ST_HTAgeToSI}} and \code{\link{ST_YrsToBH}}
}
\author{
Yong Luo
}
