% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callSiteTools.R
\docType{methods}
\name{ST_HTAgeToSI}
\alias{ST_HTAgeToSI}
\alias{ST_HTAgeToSI,numeric,numeric,numeric,numeric,numeric,character,character-method}
\alias{ST_HTAgeToSI,numeric,numeric,numeric,numeric,missing,character,missing-method}
\alias{ST_HTAgeToSI,numeric,numeric,numeric,numeric,numeric,character,missing-method}
\title{Calculate site index in site tools}
\usage{
ST_HTAgeToSI(curveRef, boredAge, ageType, height, estimateMethod,
  siteToolsDLLPath, sasExePath)


  \S4method{ST_HTAgeToSI}{numeric,numeric,numeric,numeric,numeric,character,character}(curveRef,
  boredAge, ageType, height, estimateMethod, siteToolsDLLPath, sasExePath)


  \S4method{ST_HTAgeToSI}{numeric,numeric,numeric,numeric,missing,character,missing}(curveRef,
  boredAge, ageType, height, siteToolsDLLPath)


  \S4method{ST_HTAgeToSI}{numeric,numeric,numeric,numeric,numeric,character,missing}(curveRef,
  boredAge, ageType, height, estimateMethod, siteToolsDLLPath)
}
\arguments{
\item{curveRef}{numeric, Either site index curve reference or growth intercept curve reference}

\item{boredAge}{numeric, Tree age at the bored height}

\item{ageType}{numeric, Must be either \code{0} or \code{1}. \code{0} stands for total age, for which site index is 
calculated for 50 years of total tree age. While \code{1} stands for breast height age, for which 
site index is calculated for 50 year old at breast height.}

\item{height}{numeric, Tree height}

\item{estimateMethod}{numeric, Defines how the site tools estimate site index. Valued as \code{0} and \code{1}, 
\code{0} is interative and while \code{1} is directive. Default is \code{1}, which is directive.}

\item{siteToolsDLLPath}{character, Path to \code{SINDEX33.DLL}}

\item{sasExePath}{character, Path to sas executable, i.e., \code{sas.exe}. If missing, the function takes
\code{C:/Program Files/SASHome/x86/SASFoundation/9.3} as default. However, it will cause crush
if sas executable does not located in default folder.}
}
\value{
A list of \code{output} and \code{error}. \code{output} is the site index. 
        \code{error} the flag in calculation, with a negative value represents failure.
}
\description{
Calculate site index based on bored age (\code{boredAge}), height (\code{height}) and curve reference (\code{curveRef})
             in site tools. The site index is defined as height at 50 years old. Curve reference can be either site index curve 
             reference or growth intercept curve reference. Site
             index curve reference can be derived using \code{ST_DefCurve} function, while growth intercept curve reference can be derived
             using \code{ST_DefGICurve} function.
}
\seealso{
\code{\link{ST_SpecRemap}}, \code{\link{ST_DefCurve}}, \code{\link{ST_DefGICurve}} and \code{\link{ST_YrsToBH}}
}
\author{
Yong Luo
}
