% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callSiteTools.R
\docType{methods}
\name{ST_SpecRemap}
\alias{ST_SpecRemap}
\alias{ST_SpecRemap,character,character,character,character-method}
\alias{ST_SpecRemap,character,character,character,missing-method}
\title{Calculate site index reference by species and region}
\usage{
ST_SpecRemap(species, ICRegion, siteToolsDLLPath, sasExePath)

\S4method{ST_SpecRemap}{character,character,character,character}(species,
  ICRegion, siteToolsDLLPath, sasExePath)

\S4method{ST_SpecRemap}{character,character,character,missing}(species,
  ICRegion, siteToolsDLLPath)
}
\arguments{
\item{species}{character, Species code, must be consistent with the species code in site tools.}

\item{ICRegion}{character, Must be either \code{I} (interior) and \code{C} (coastal). In VRI compiler,
IC regions are derived using \code{\link{BEC2IC}} function.}

\item{siteToolsDLLPath}{character, Path to \code{SINDEX33.DLL}}

\item{sasExePath}{character, Path to sas executable, i.e., \code{sas.exe}. If missing, the function takes
\code{C:/Program Files/SASHome/x86/SASFoundation/9.3} as default. However, it will cause crush
if sas executable is not located in default path.}
}
\value{
Site index reference that can be recognized by sitetools
}
\description{
This function is to extract site index reference for a given species and region in siteTools program.
}
\seealso{
\code{\link{ST_DefCurve}}, \code{\link{ST_DefGICurve}}, \code{\link{ST_HTAgeToSI}} and \code{\link{ST_YrsToBH}}
}
\author{
Yong Luo
}
