% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callSiteTools.R
\docType{methods}
\name{ST_YrsToBH}
\alias{ST_YrsToBH}
\alias{ST_YrsToBH,numeric,numeric,character,character-method}
\alias{ST_YrsToBH,numeric,numeric,character,missing-method}
\title{Derive years between ground and breast height in site tools}
\usage{
ST_YrsToBH(SICurveRef, siteIndex, siteToolsDLLPath, sasExePath)

\S4method{ST_YrsToBH}{numeric,numeric,character,character}(SICurveRef,
  siteIndex, siteToolsDLLPath, sasExePath)

\S4method{ST_YrsToBH}{numeric,numeric,character,missing}(SICurveRef, siteIndex,
  siteToolsDLLPath)
}
\arguments{
\item{SICurveRef}{numeric, Site index curve reference in site tools. It can be derived using \code{ST_DefCurve} function.}

\item{siteIndex}{numeric, Site index. Defined as tree height at 50 years old.}

\item{siteToolsDLLPath}{character, Path to \code{SINDEX33.DLL}}

\item{sasExePath}{character, Path to sas executable, i.e., \code{sas.exe}. If missing, the function takes
\code{C:/Program Files/SASHome/x86/SASFoundation/9.3} as default. However, it will cause crush
if sas executable does not located in default folder.}
}
\value{
A list of \code{output} and \code{error}. \code{output} is the derived years to breast height. 
       \code{error} is the flag in calculation, with negative value represents failure.
}
\description{
Derive years between ground and breast height based on site index curve reference (\code{SICurveRef}) and 
             site index (\code{siteIndex}) in site tools.
}
\seealso{
\code{\link{ST_SpecRemap}}, \code{\link{ST_DefCurve}}, \code{\link{ST_DefGICurve}} and \code{\link{ST_HTAgeToSI}}
}
\author{
Yong Luo
}
