% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toWSVRatio.R
\docType{methods}
\name{toWSVRatio}
\alias{toWSVRatio}
\alias{toWSVRatio,data.table,numeric,numeric-method}
\alias{toWSVRatio,data.table,missing,numeric-method}
\alias{toWSVRatio,data.table,numeric,missing-method}
\alias{toWSVRatio,data.table,missing,missing-method}
\title{Calculate to whole stem volume ratios}
\usage{
toWSVRatio(ratioData, minDBH, minObs)

\S4method{toWSVRatio}{data.table,numeric,numeric}(ratioData, minDBH, minObs)

\S4method{toWSVRatio}{data.table,missing,numeric}(ratioData, minObs)

\S4method{toWSVRatio}{data.table,numeric,missing}(ratioData, minDBH)

\S4method{toWSVRatio}{data.table,missing,missing}(ratioData)
}
\arguments{
\item{ratioData}{data.table, The data used to calculated for deriving toWSV ratios.}

\item{minDBH}{numeric, Defines minimum DBH threshold to select trees in deriving the ratios. If missing 10 cm
is used.}

\item{minObs}{numeric, Defined minimum of observations that for each conbination that used in deriving ratios.
If missing, 3 is used.}
}
\value{
A ratio table
}
\description{
This function calculates a bunch of ratio to adjust volume components based on whole stem volume \code{WSV}.
             For each combination of project group, live and dead status and species. Specifically,
             the function calculates the ratios of whole stem volume (\code{WSV}), net stem volume (\code{NET}), 
             merchantable volume (\code{MER}), net merchantable volume (\code{NETM}), 
             net volume after waste 2 (\code{NTW2}), 
             volume of less top, stemp, crurser decay, waste and breakage (\code{NTWB}),
             volume of less top, stump and decay (\code{D}), volume of less top, stump, waste and decay (\code{DW}),
             volume of less top, stump, waste, decay and breakage (\code{DWB})
             to whole stem volume. The ratios are calculated as the mean of tree with DBH >= 10.
             The function is equivalent to \code{calc_ratio_2017.sas}.
}
\author{
Yong Luo
}
