% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTM_Convertor.R
\docType{methods}
\name{UTM_Convertor}
\alias{UTM_Convertor}
\alias{UTM_Convertor,character,integer,integer,integer,character,character-method}
\alias{UTM_Convertor,character,integer,integer,integer,missing,character-method}
\alias{UTM_Convertor,character,integer,integer,integer,character,missing-method}
\alias{UTM_Convertor,character,integer,integer,integer,missing,missing-method}
\alias{UTM_Convertor,missing,integer,integer,integer,missing,missing-method}
\alias{UTM_Convertor,missing,integer,integer,integer,missing,character-method}
\title{Convert UTM to other coordinate reference system.}
\usage{
UTM_Convertor(point_ID, zone, northing, easting, CRS_To, class)


  \S4method{UTM_Convertor}{character,integer,integer,integer,character,character}(point_ID,
  zone, northing, easting, CRS_To, class = "character")


  \S4method{UTM_Convertor}{character,integer,integer,integer,missing,character}(point_ID,
  zone, northing, easting, class)


  \S4method{UTM_Convertor}{character,integer,integer,integer,character,missing}(point_ID,
  zone, northing, easting, CRS_To)


  \S4method{UTM_Convertor}{character,integer,integer,integer,missing,missing}(point_ID,
  zone, northing, easting)

\S4method{UTM_Convertor}{missing,integer,integer,integer,missing,missing}(zone,
  northing, easting)


  \S4method{UTM_Convertor}{missing,integer,integer,integer,missing,character}(zone,
  northing, easting, class)
}
\arguments{
\item{point_ID}{character, Data point ID.}

\item{zone}{integer, UTM zone.}

\item{northing}{integer, UTM northing.}

\item{easting}{integer, UTM easting.}

\item{CRS_To}{character, Defines the spatial coordination reference that you wish to transform.
Default is BC Albers reference system.}

\item{class}{character, Define the class of returned objective. Currently this function supports
either \code{table} or \code{sp} class. Default is \code{table}.}
}
\value{
Reprojected objective.
}
\description{
Converts UTM coordinates to the other coordinate reference system.
}
\author{
Yong Luo
}
