% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VRICompiler.R
\docType{methods}
\name{VRICompiler}
\alias{VRICompiler}
\title{VRI compiler - VRI specific}
\usage{
VRICompiler(dataSourcePath, outputPath = ".", equation = "KBEC",
  walkThru = TRUE, logMinLength = 0.1, stumpHeight = 0.3,
  breastHeight = 1.3, UTOPDIB = 10, utilLevel = 4, weirdUtil = "No",
  siteToolsDLLPath,
  sasExePath = "C:/Program Files/SASHome/x86/SASFoundation/9.3")
}
\arguments{
\item{dataSourcePath}{character, Specifies the path that directs to the VRI original data soruce.}

\item{outputPath}{character, Specifies the folder to save all the outputs.}

\item{equation}{character, Specifies the taper equation that is used for compiler. Currently supports
BEC-based (\code{KBEC}) and FIZ-based (\code{KFIZ}).}

\item{walkThru}{logical, Speciefies whether the data had been collected using work through method. Default is \code{TRUE},
if it is not specified.}

\item{logMinLength}{numeric, Specifies minimum length of log when doing log length adjustment,
see \code{\link{logMatrixAdjustment}} for details. If missing 0.1 is used.}

\item{stumpHeight}{numeric, Stump height. If missing 0.3 is used.}

\item{breastHeight}{numeric, Breast height. If missing 1.3 is used.}

\item{UTOPDIB}{numeric, Threshold inside-bark diameter for merchantable volume. If missing, UTOPDIB is 10.}

\item{utilLevel}{numeric, Specifies utilization level in summrizing tree volumes at cluster and species level. Default is 4.}

\item{weirdUtil}{character, Specifies weird utilization in summarizing tree volumes at cluster and species level.
Default is \code{no}, if missing. Otherwise, a number should be provided.}

\item{siteToolsDLLPath}{character, Path to SINDEX33.DLL.}

\item{sasExePath}{character, Path to sas executable, i.e., sas.exe. If missing,
the function takes \code{C:/Program Files/SASHome/x86/SASFoundation/9.3 as default}.
However, it will cause crush if sas executable is not located in default path.}
}
\value{
This function compiles data and save outputs in \code{outputPath} and no file is returned.
}
\description{
This function compiles VRI data by calling specific VRI functions. Unlike the original
             compiler (i.e., SAS compiler), the R version compiler hardcodes all the lookup tables in
             the compilation process. Please refer the descriptions for lookup table to see whether
             they are same as the original lookup table.
}
\references{
VRI compiler manual
}
\author{
Yong Luo
}
