% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSV_BARegression.R
\docType{methods}
\name{WSV_BARegression}
\alias{WSV_BARegression}
\alias{WSV_BARegression,data.table,numeric,numeric-method}
\alias{WSV_BARegression,data.table,missing,numeric-method}
\alias{WSV_BARegression,data.table,numeric,missing-method}
\alias{WSV_BARegression,data.table,missing,missing-method}
\title{Create regression equation to link whole stem volume and basal area in VRI data-VRI specific}
\usage{
WSV_BARegression(regressionData, minObs, minR2)

\S4method{WSV_BARegression}{data.table,numeric,numeric}(regressionData, minObs,
  minR2)

\S4method{WSV_BARegression}{data.table,missing,numeric}(regressionData, minR2)

\S4method{WSV_BARegression}{data.table,numeric,missing}(regressionData, minObs)

\S4method{WSV_BARegression}{data.table,missing,missing}(regressionData)
}
\arguments{
\item{regressionData}{numeric, Specifies whole stem volume.}

\item{minObs}{numeric, Defines the minimum number of observations (included) that used for the regression.
If missing, 3 is used to be consistent with the compilation manual.}

\item{minR2}{numeric, Defines the minimum R square to identify the models goodness of fit.
If missing, 0.3 is used.}
}
\value{
A data table that contains key statistics of whole stem volume and basal area regression.
}
\description{
This function develops regression equations between whole stem volume (\code{WSV}) and basal area (\code{BA})
             for both alive and dead trees. The equation is derived at project group and species level. When sample size is small 
             or the model's performance is poor for a given species, the species type, i.e., deciduous or coniferious species, is
             used as this species in developing the model. The function is equivalent to \code{reg_wsv_2017.sas}.
             To estimate tree volume for non-enhanced trees, the resultant equations will be used.
}
\author{
Yong Luo
}
